# backend/tests/api/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ExecuteAgentRequest": {
        "description": "Optional body for ``POST /workspaces/{id}/agents/{type}/execute``.\n\nPRD-0006 Phase 2 / IMPL-0007 \u00a7B4 \u2014 adds ``user_note`` for the planner's\nRefine flow. Other agent types accept the body but ignore the note.",
        "properties": {
          "user_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Note"
          }
        },
        "title": "ExecuteAgentRequest",
        "type": "object"
      },
      "ExecuteResponse": {
        "properties": {
          "agent_run_id": {
            "title": "Agent Run Id",
            "type": "string"
          },
          "agent_type": {
            "title": "Agent Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "agent_run_id",
          "agent_type",
          "status"
        ],
        "title": "ExecuteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PermissionDecision": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          }
        },
        "required": [
          "approved"
        ],
        "title": "PermissionDecision",
        "type": "object"
      },
...
      "RunAllResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "RunAllResponse",
        "type": "object"
      },
...
      "SuggestNextResponse": {
        "properties": {
          "action_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Type"
          },
          "agent_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Type"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "agent_type",
          "reason",
          "priority"
        ],
        "title": "SuggestNextResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspaces/{workspace_id}/agent-execution/stream": {
      "get": {
        "description": "Stream permission_request and done events during agent execution.\n\nThe frontend connects to this while an agent is running. Events:\n- permission_request: agent needs user approval for a tool\n- done: agent execution has completed (success or failure)\n\nIf the client disconnects while a permission is pending, the pending\napproval is auto-denied to unblock the executor.",
        "operationId": "stream_agent_execution_api_workspaces__workspace_id__agent_execution_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Agent Execution",
        "tags": [
          "agent-execution"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/agent-runs/{run_id}/cancel": {
      "post": {
        "description": "Cancel a running agent run.",
        "operationId": "cancel_agent_run_api_workspaces__workspace_id__agent_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Agent Run",
        "tags": [
          "agent-execution"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agent-runs/{run_id}/permission": {
      "post": {
        "description": "Approve or deny a pending tool-use permission request.",
        "operationId": "respond_to_permission_api_workspaces__workspace_id__agent_runs__run_id__permission_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Respond To Permission",
        "tags": [
          "agent-execution"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agents/{agent_type}/execute": {
      "post": {
        "description": "Start an agent run as a background task.\n\nReturns immediately with the agent_run_id. Connect to the\nagent-execution SSE stream to receive permission_request events\nand a done signal.\n\nOptional body ``{user_note}`` is forwarded to the planner's prompt for\nPRD-0006 Phase 2's Refine flow; other agent types ignore it.",
        "operationId": "execute_agent_api_workspaces__workspace_id__agents__agent_type__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_type",
            "required": true,
            "schema": {
              "title": "Agent Type",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ExecuteAgentRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Agent",
        "tags": [
          "agent-execution"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/pipeline/run-all": {
      "post": {
        "description": "Run all remaining agents in pipeline order as a background task.\n\nEach agent runs sequentially. Progress events stream via the\nagent-execution SSE endpoint. Stops on first failure.",
        "operationId": "run_all_pipeline_api_workspaces__workspace_id__pipeline_run_all_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAllResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run All Pipeline",
        "tags": [
          "agent-execution"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/pipeline/suggest-next": {
      "get": {
        "description": "Return the recommended next agent based on current context state.",
        "operationId": "suggest_next_endpoint_api_workspaces__workspace_id__pipeline_suggest_next_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestNextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest Next Endpoint",
        "tags": [
          "agent-execution"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/plan/approve": {
      "post": {
        "description": "Mark the workspace's remediation plan as approved.\n\nPRD-0006 Story 3 \u2014 the planner pauses the run-all loop until the user\nexplicitly approves. This endpoint flips ``plan.approved=true`` in BOTH\nstores: the SQLite sidebar (read by the Issues-page derivation) AND\nthe filesystem ``context/plan.json`` (read by ``suggest_next`` to\ndecide whether the executor may run). A subsequent\n``POST /pipeline/run-all`` will then suggest the executor.\n\nReturns 404 if the workspace doesn't exist or the planner hasn't yet\nwritten a plan to the sidebar.",
        "operationId": "approve_plan_api_workspaces__workspace_id__plan_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SidebarState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Plan",
        "tags": [
          "agent-execution"
        ]
      }
    },
...
  }
}